Fixed mistake in readme about pageview arguments
Looking at the source, `pageview` has also a session duration argument that is missed in the readme. Adding it fixed my problem: I couldn't get the clientID to work because it was not in the right argument place.

I used `undefined` without any issues. Can you please explain what is the right thing to put for the session duration?

(also clarified how `clientID` is found)